Add clean patterm
