+++ app/src/main/java/com/org/marton/studio/project/moviesappcompose/data/DetailMovieResponse.kt
@@ -1,0 +1,19 @@
+package com.org.marton.studio.project.moviesappcompose.data
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class DetailMovieResponse(
+    @SerialName("id") val id: Int,
+    @SerialName("title") val title: String,
+    @SerialName("adult") val adult: Boolean,
+    @SerialName("overview") val overview: String,
+    @SerialName("release_date") val releaseDate: String,
+    @SerialName("poster_path") val posterPath: String?,
+    @SerialName("backdrop_path") val backdropPath: String?,
+    @SerialName("original_title") val originalTitle: String,
+    @SerialName("original_language") val originalLanguage: String,
+    @SerialName("popularity") val popularity: Double,
+    @SerialName("vote_average") val voteAverage: Double,
+)
