+++ app/src/main/java/com/org/marton/studio/project/moviesappcompose/data/MovieResponse.kt
@@ -11,17 +11,3 @@ data class MovieResponse(
     @SerialName("total_results") val totalResults: Int? = null
 )
 
-@Serializable
-data class DetailMovieResponse(
-    @SerialName("id") val id: Int,
-    @SerialName("title") val title: String,
-    val adult: Boolean,
-    @SerialName("overview") val overview: String,
-    @SerialName("release_date") val releaseDate: String,
-    @SerialName("poster_path") val posterPath: String?,
-    @SerialName("backdrop_path") val backdropPath: String?,
-    @SerialName("original_title") val originalTitle: String,
-    @SerialName("original_language") val originalLanguage: String,
-    @SerialName("popularity") val popularity: Double,
-    @SerialName("vote_average") val voteAverage: Double,
-)
