--- app/src/main/java/com/org/marton/studio/project/moviesappcompose/data/MoviesRepository.kt
@@ -1,29 +1,0 @@
-package com.org.marton.studio.project.moviesappcompose.data
-
-import com.org.marton.studio.project.moviesappcompose.ui.theme.Movie
-
-class MoviesRepository(private val moviesService: MovieService) {
-    suspend fun fetchPopularMovies(): List<Movie> {
-        return moviesService.getPopularMovies().results?.map { it.toDomainMovie() } ?: emptyList()
-    }
-
-    suspend fun fetchMovieDetails(id: String): Movie {
-        return moviesService.fetchMovieDetails(id).toDomainMovie()
-    }
-
-}
-
-fun DetailMovieResponse.toDomainMovie() = Movie(
-    id = id,
-    title = title,
-    poster = if (posterPath != null) "https://image.tmdb.org/t/p/w185/$posterPath" else "",
-    overview = overview,
-    releaseNote = releaseDate,
-    backdrop = backdropPath.let { "https://image.tmdb.org/t/p/w780/$it" },
-    originalTitle = originalTitle,
-    originalLanguage = originalLanguage,
-    popularity = popularity,
-    voteAverage = voteAverage,
-)
-
-
