+++ app/src/main/java/com/org/marton/studio/project/moviesappcompose/domain/MoviesRepository.kt
@@ -1,0 +1,6 @@
+package com.org.marton.studio.project.moviesappcompose.domain
+
+interface MoviesRepository {
+    suspend fun fetchPopularMovies(): List<Movie>
+    suspend fun fetchMovieDetails(id: String): Movie
+}
